requirements.txt
requests
exa_py
python-dotenv
pandas
